// src/domain/types/role.ts
export type Role =
  | "RECEPTIONIST"
  | "RESERVATION_STAFF"
  | "CASHIER"
  | "FRONT_OFFICE_SUPERVISOR"
  | "HOUSEKEEPING"
  | "HOUSEKEEPING_SUPERVISOR"
  | "ENGINEERING"
  | "HOTEL_MANAGER"
  | "FINANCE_HOTEL"
  | "HEAD_OFFICE_APPROVER"
  | "HEAD_OFFICE_FINANCE"
  | "AUDITOR"
  | "SYSTEM_ADMIN"
  | "OWNER";